CSS Media Queries
There's a New & easy ways to use Media Queries:

@media only screen and (width <= 360px) {
        element{
            // code //
        }
      }

We can use other operators instead of writing (min-width) or (max-width)

= equals to
> greater than [your size]
< smaller than [ your size]

and so on ...
<=
>=
[ CSS 2 --> Media Rule ] then in [ CSS 3 --> Media Tool ] ::::: - For the purpose of making responsive Websites, 'Responsive' in the sense of it's ability to adapt according to the different sizes of devices, the website is being accessed through.

SYNTAX :-

@media screen and (property : ) { css code }
@media only screen and (property : ) { css code }
@media not screen and (property : ){ css code }

Here,
property can be (max-width : 12px), (min-width : ), (orientation : )

CSS media queries are a way to apply different styles based on characteristics of the user's device or browser. They allow you to create responsive designs that adapt to different screen sizes, resolutions, and other device properties.
Media queries are typically used in combination with the @media rule. Here's an example of how you might use a media query in CSS:
cssCopy code
@media screen and (max-width: 600px) { body { background-color: lightblue; } }
In this example, the styles inside the curly braces will only apply if the conditions specified in the media query are met. In this case, the styles will apply if the screen width is 600 pixels or less.
Here's a breakdown of the components of the media query:
•	@media: This is the starting point of the media query.
•	screen: This specifies the type of media being targeted. In this case, it's targeting screens (like those of computers and mobile devices).
•	and: This is a logical operator that allows you to combine multiple conditions.
•	(max-width: 600px): This is the actual condition being tested. In this case, it checks if the maximum width of the viewport is 600 pixels or less.
You can use various properties and values in media queries to target different characteristics. Some common ones include:
•	width: Specifies the width of the viewport.
•	height: Specifies the height of the viewport.
•	min-width and max-width: Define a range of widths.
•	min-height and max-height: Define a range of heights.
•	orientation: Checks if the device is in portrait or landscape mode.
•	aspect-ratio: Compares the width and height of the viewport.
Here's an example of a media query using min-width and max-width:
cssCopy code
@media screen and (min-width: 768px) and (max-width: 1024px) { /* Styles for devices with widths between 768px and 1024px */ }
Using media queries, you can create styles that adapt to various device sizes, providing a better user experience across different screens and devices. This is a key aspect of responsive web design.