CSS Float and Clear
What is the essence of this video:
1. float: left; or float: right; => provides display-flex-like properties. but can be controlled by its position interchanging the value left and right.
2. If the content is bigger than the size of the other divs then the content will overflow.
3. To prevent the overflow, simply use the CSS property display: flow-root;
4. To avoid an item from overlapping with another div simply use the CSS property clear: left, clear: right, or clear: both; as per the requirement.
float and clear are two CSS properties used for controlling the layout and positioning of elements.
1.	float:
The float property is used to specify that an element should be taken out of the normal flow and placed along the left or right side of its container. This is commonly used for creating layouts with multiple columns or for positioning elements next to each other.
For example, if you want to float an element to the left, you can do:
cssCopy code
.example { float: left; }
Similarly, you can use float: right; to float an element to the right.
It's important to note that when an element is floated, it may affect the layout of surrounding elements. This is where the clear property comes into play.
2.	clear:
The clear property is used to control the behavior of elements that are next to a floated element. It determines whether an element should be moved below (cleared) the floated elements.
There are different values for clear:
•	clear: left;: The element will not allow floating elements on its left side.
•	clear: right;: The element will not allow floating elements on its right side.
•	clear: both;: The element will not allow floating elements on either side (left or right).
For example, if you have a floated element and you want to make sure the next element appears below it, you can use:
cssCopy code
.next-element { clear: both; }
This will ensure that .next-element is positioned below any floated elements.
Here's an example of how float and clear can be used together:
cssCopy code
.left-box { float: left; width: 50%; background-color: #3498db; } .right-box { float: right; width: 50%; background-color: #e74c3c; } .clear-element { clear: both; /* This element will appear below the floated elements */ }
htmlCopy code
<div class="left-box">Left Box</div> <div class="right-box">Right Box</div> <div class="clear-element">Clear Element</div>
In this example, .left-box and .right-box are floated to the left and right respectively. The .clear-element ensures that it appears below both floated elements due to clear: both;.